src/requirements.txt
# 依存パッケージ
mecab-python3==1.0.4
unidic-lite==1.0.8
pyyaml==6.0
pydantic==2.5.0
pandas==2.1.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
numpy>=1.24.0
regex
tqdm
pytest
fugashi
protobuf